Move modal to shared components
