+++ ui/packages/components/src/Modal/index.ts
@@ -1,0 +1,1 @@
+export { Modal } from './Modal';
